+++ src/capturer/simple.rs
@@ -125,14 +125,14 @@ impl Monitor {
 
 #[cfg(test)]
 mod tests {
-  use crate::{capturer::model::Capturer, utils::FrameInfoExt, Monitor};
+  use crate::{capturer::model::Capturer, utils::FrameInfoExt, Scanner};
   use serial_test::serial;
   use std::{thread, time::Duration};
 
   #[test]
   #[serial]
   fn simple_capturer() {
-    let ctx = Monitor::factory().unwrap().next().unwrap();
+    let ctx = Scanner::new().unwrap().next().unwrap();
     let mut capturer = ctx.simple_capturer().unwrap();
 
     // sleep for a while before capture to wait system to update the screen