+++ src/lib.rs
@@ -1,5 +1,6 @@
 mod error;
 mod monitor;
+mod scanner;
 
 pub mod capturer;
 pub mod model;
@@ -7,3 +8,4 @@ pub mod utils;
 
 pub use error::*;
 pub use monitor::*;
+pub use scanner::*;